requirements.txt
mcp>=1.0.0
chromadb>=0.4.0
langchain-text-splitters>=0.0.1
beautifulsoup4>=4.12.0
lxml>=5.0.0
requests>=2.31.0
gitpython>=3.1.0
python-docx>=1.0.0
openpyxl>=3.1.0
click>=8.1.0
rich>=13.0.0
python-dotenv>=1.0.0
python-frontmatter>=1.0.0
